Renamed Minicloner.Tests.Fakes.AccessModifiers namespace to Minicloner.Tests.Fakes.Fields Closes #28
