+++ src/Minicloner/Minicloner.Tests/Fakes/Fields/Class_With_ProtectedInternalField.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Minicloner.Tests.Fakes.AccessModifiers
+namespace Minicloner.Tests.Fakes.Fields
 {
     public class Class_With_ProtectedInternalField
     {
